+++ server/Player.h
@@ -9,12 +9,12 @@ public:
 	Player(int playerId, Connection* connection, char* name);
 	~Player();
 
-	void SetPos(int x, int y, int z);
+	void SetPos(float x, float y, float z);
 
 public:
 	int playerId;
 	char* name;
-	int x, y, z;
+	float x, y, z;
 	Connection* connection;
 
 private:
